requirements.txt
pydantic>=2.0.0
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0